backend/requirements.txt
amqp==5.3.1
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
bcrypt==3.2.0
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==7.1.2
click-didyoumean==0.3.1
click-repl==0.3.0
cookiecutter==2.1.1
cookiecutter-data-science==2.2.0
fastapi==0.65.2
FastAPI-Cookiecutter==0.1.2
ghp-import==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==2.10
iniconfig==2.1.0
Jinja2==3.1.6
jinja2-time==0.2.0
kombu==5.5.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
packaging==25.0
passlib==1.7.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
py==1.11.0
pycparser==2.22
pydantic==1.10.22
Pygments==2.19.1
PyJWT==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.5
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.0
qrcode==8.2
requests==2.28.2
rich==14.0.0
sniffio==1.3.1
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.16
uvicorn==0.12.1
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
